requirements.txt
Flask
python-dotenv
azure-identity
openai
requests
gunicorn